chore: extract load-broadcast-replay suite into src/suites/load-broadcast-replay.ts
Part of REAL-1037.
