+++ test/e2e/realtime-check.ts
@@ -20,6 +20,7 @@ import { setup, cleanup } from "./src/fixtures.ts";
 import { broadcastBinary } from "./src/suites/broadcast-binary.ts";
 import { authorization } from "./src/suites/authorization.ts";
 import { loadBroadcastFromDb } from "./src/suites/load-broadcast-from-db.ts";
+import { loadBroadcastReplay } from "./src/suites/load-broadcast-replay.ts";
 
 async function runConnectionTest() {
   suite("connection");
@@ -297,40 +298,6 @@ async function runLoadBroadcastTests() {
       await settle(() => latencies.length, LOAD_MESSAGES, LOAD_SETTLE_MS);
 
       return measureThroughput(latencies, LOAD_MESSAGES, "broadcast API events", LOAD_DELIVERY_SLO);
-    } finally {
-      await stopClient(supabase);
-    }
-  });
-}
-
-async function runLoadBroadcastReplayTests(testUser: { email: string; password: string }) {
-  suite("load-broadcast-replay");
-
-  await sleep(RATE_LIMIT_PAUSE_MS);
-  await test("broadcast replay throughput", async () => {
-    const supabase = createClient(PROJECT_URL, ANON_KEY, { realtime: REALTIME_OPTS });
-    try {
-      await signInUser(supabase, testUser.email, testUser.password);
-      const event = crypto.randomUUID();
-      const topic = randomTopic();
-
-      const since = Date.now() - 1000;
-      await Promise.all(Array.from({ length: LOAD_MESSAGES }, (_, i) =>
-        supabase.from("replay_check").insert({ id: crypto.randomUUID(), topic, event, payload: { seq: i } })
-      ));
-
-      const latencies: number[] = [];
-      const replayStart = performance.now();
-      const receiver = supabase.channel(topic, {
-        config: { private: true, broadcast: { replay: { since, limit: 25 } } },
-      }).on("broadcast", { event }, () => {
-        latencies.push(performance.now() - replayStart);
-      });
-      await openChannel(receiver);
-
-      await settle(() => latencies.length, LOAD_MESSAGES, LOAD_SETTLE_MS);
-
-      return measureThroughput(latencies, LOAD_MESSAGES, "replayed broadcast events", LOAD_DELIVERY_SLO);
     } finally {
       await stopClient(supabase);
     }
@@ -1376,7 +1343,7 @@ const descriptors: SuiteDescriptor[] = [
   { name: "load-presence", label: "load-presence", needsDb: false, run: () => runLoadPresenceTests() },
   { name: "load-broadcast", label: "load-broadcast", needsDb: false, run: () => runLoadBroadcastTests() },
   loadBroadcastFromDb,
-  { name: "load-broadcast-replay", label: "load-broadcast-replay", needsDb: true, run: ({ testUser }) => runLoadBroadcastReplayTests(testUser) },
+  loadBroadcastReplay,
   { name: "broadcast", label: "broadcast extension", needsDb: false, run: () => runBroadcastTests() },
   { name: "broadcast-replay", label: "broadcast replay", needsDb: true, run: ({ testUser, supabase }) => runBroadcastReplayTests(testUser, supabase) },
   { name: "presence", label: "presence extension", needsDb: true, run: ({ testUser, supabase }) => runPresenceTests(testUser, supabase) },
